requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dataclasses-json==0.6.7
distro==1.9.0
docarray==0.40.0
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
jiter==0.6.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.1
langchain-community==0.3.1
langchain-core==0.3.7
langchain-openai==0.2.1
langchain-pinecone==0.1.3
langchain-text-splitters==0.3.0
langsmith==0.1.134
markdown-it-py==3.0.0
MarkupSafe==3.0.1
marshmallow==3.22.0
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.9.3
numpy==1.26.4
openai==1.51.0
orjson==3.10.7
packaging==24.1
pandas==2.2.3
pillow==10.4.0
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
propcache==0.2.0
protobuf==5.28.2
pyarrow==17.0.0
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytube==15.0.0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.2
rpds-py==0.20.0
ruff==0.6.9
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.35
streamlit==1.39.0
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.8.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
types-requests==2.32.0.20240914
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
yarl==1.15.0
